requirements.txt
allure-behave=2.13.5
allure-pytest=2.13.5
allure-python-commons=2.13.5
attrs=24.3.0
behave=1.2.6
colorama=0.4.6
iniconfig=2.0.0
packaging=24.2
parse=1.20.2
parse_type=0.6.4
pip=24.2
pluggy=1.5.0
pytest=8.3.4
six=1.17.0